% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea.R
\name{new_nmea}
\alias{new_nmea}
\alias{validate_nmea}
\title{S3 Destails for the 'nmea' class}
\usage{
new_nmea(x = list())

validate_nmea(x)
}
\arguments{
\item{x}{A \code{list()} of \code{raw()} or \code{NULL}.}
}
\value{
An object of class 'nmea'.
}
\description{
S3 Destails for the 'nmea' class
}
