% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea.R
\name{nmea}
\alias{nmea}
\alias{as_nmea}
\alias{as_nmea.nmea}
\alias{as_nmea.list}
\alias{as_nmea.data.frame}
\alias{as_nmea.raw}
\alias{as_nmea.character}
\title{Create NMEA sentences}
\usage{
nmea(x = list())

as_nmea(x, ...)

\method{as_nmea}{nmea}(x, ...)

\method{as_nmea}{list}(x, ...)

\method{as_nmea}{data.frame}(x, ..., nmea_col = "sentence")

\method{as_nmea}{raw}(x, ...)

\method{as_nmea}{character}(x, ...)
}
\arguments{
\item{x}{An object to be converted to one or more NMEA sentences}

\item{...}{Unused}

\item{nmea_col}{A column name containing the NMEA sentence in a data.frame.
Defaults to "sentence" as this is the column returned by \code{\link[=read_nmea]{read_nmea()}}.}
}
\value{
A \code{\link[=new_nmea]{new_nmea()}}.
}
\description{
Most NMEA sentences are valid ASCII text. Some, however, contain embedded
nul characters as part of their structure and can't be represented by
a \code{character()}. For most NMEA text you probably don't need this class
(just read the file as lines and pass to the other functions in
this package).
}
\examples{
nmea()
as_nmea(NA_character_)
as_nmea(nmea_test_basic)
as_nmea(list(charToRaw(nmea_test_basic[1])))
as_nmea(charToRaw(nmea_test_basic[1]))

# can include 0x00!
nmea(list(as.raw(0x00)))

}
