% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmearead.R
\name{nmea_calculateaveragesnr}
\alias{nmea_calculateaveragesnr}
\title{Calculate average SNR}
\usage{
nmea_calculateaveragesnr(gpstable, sortedsatellites = NULL)
}
\arguments{
\item{gpstable}{The data.table containing the GPS details as returned by \code{\link{nmea_readfiles}}
or \code{\link{nmea_parsetext}}.}

\item{sortedsatellites}{If an integer value is provided, it will sort the satellites it 
uses for determining the position by SNR descending, and calculate the average SNR of the top provided amount.}
}
\value{
Vector with average SNR. This vector can be appended as a new column using \code{\link[data.table]{cbind}}
to the provided \code{gpstable}.
}
\description{
Calculate the average SNR for all or a certain amount of satellites.
}
