% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-spec.R
\name{nmea_col_character}
\alias{nmea_col_character}
\alias{nmea_col_double}
\alias{nmea_col_skip}
\alias{nmea_col_blob}
\title{Specify NMEA field formats}
\usage{
nmea_col_character()

nmea_col_double()

nmea_col_skip()

nmea_col_blob(col_end = ",", col_end_sep = col_end)
}
\arguments{
\item{col_end}{A sequence of characters to search for
at the end of a sequence. Usually this is the same as
\code{col_end_sep} but for blob fields you probably need to specify
more than one character.}

\item{col_end_sep}{The last character that should be skipped until
before the next field can be parsed. This is almost always a comma
(",").}
}
\value{
An object of class 'nmea_col'
}
\description{
Specify NMEA field formats
}
