% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-util.R
\name{nmea_date_time}
\alias{nmea_date_time}
\alias{nmea_longitude}
\alias{nmea_latitude}
\title{NMEA utility functions}
\usage{
nmea_date_time(date, time)

nmea_longitude(degree_spec, hemisphere)

nmea_latitude(degree_spec, hemisphere)
}
\arguments{
\item{date}{A Date object or NMEA datestamp (DDMMYY)}

\item{time}{A hms object or NMEA time (HHMMSS)}

\item{degree_spec}{A longitude/latitude as it appears in the NMEA sentence
(e.g., "4916.45"). This is in the form (dddmm.mm) where there are always
two minutes digits to the left of the decimal point.}

\item{hemisphere}{One of "S", "N", "W", or "E".}
}
\value{
\itemize{
\item \code{nmea_date_time}: A POSIXct datetime (UTC)
\item \code{nmea_longitude} and \code{nmea_latitude}: A numeric longitude/latitude
}
}
\description{
Some values like datetime, longitude, and latitude, are distributed
accross fields.
}
\examples{
nmea_date_time("191194", "225446")
nmea_longitude("12311.12", "W")
nmea_latitude("4916.45", "N")

}
