% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmearead.R
\name{nmea_getpossibletypes}
\alias{nmea_getpossibletypes}
\title{Get possible types}
\usage{
nmea_getpossibletypes()
}
\value{
The string vector of the possible NMEA line types that can be processed.
}
\description{
Get a string vector of the possible NMEA line types that can be processed.
}
