% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-meta.R
\name{nmea_length}
\alias{nmea_length}
\alias{nmea_split_fields}
\alias{nmea_sub}
\alias{nmea_sentence_id}
\alias{nmea_talker}
\alias{nmea_message_type}
\alias{nmea_talker_label}
\alias{nmea_message_type_label}
\alias{nmea_checksum}
\alias{nmea_meta}
\title{Extract NMEA components}
\usage{
nmea_length(x)

nmea_split_fields(x, names = NULL, split_chars = c(",", "*"))

nmea_sub(x, start = 0L, end = -1L)

nmea_sentence_id(x)

nmea_talker(x)

nmea_message_type(x)

nmea_sentence_id(x)

nmea_talker_label(x)

nmea_message_type_label(x)

nmea_checksum(x)

nmea_meta(x)
}
\arguments{
\item{x}{An \code{\link[=nmea]{nmea()}} object.}

\item{names}{Names to assign to fields or \code{NULL} to use a noisily
assigned default.}

\item{split_chars}{A character vector of split characters that delineate
fields.}

\item{start, end}{Zero-based to extract. Negative indices refer to values
from the end of the sentence.}
}
\value{
A vector of the specified components
}
\description{
Extract NMEA components
}
\examples{
nmea_length(nmea_test_basic)
nmea_sentence_id(nmea_test_basic)
nmea_talker(nmea_test_basic)
nmea_message_type(nmea_test_basic)
nmea_talker_label(nmea_test_basic)
nmea_message_type_label(nmea_test_basic)
nmea_checksum(nmea_test_basic)

nmea_meta(nmea_test_basic)

nmea_sub(nmea_test_basic, 0, 6)
nmea_split_fields(nmea_test_basic)

}
