% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-meta.R
\name{nmea_meta}
\alias{nmea_meta}
\title{NMEA Sentence meta}
\usage{
nmea_meta(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A data.frame with columns:
\itemize{
\item \code{talker}: The two-letter talker identifier
\item \code{message_type}: A character vector of the (usually three-letter)
message type identifier
\item \code{talker_label}, \code{message_type_label}: Human-readable labels
whose value may change over time.
}
}
\description{
NMEA Sentence meta
}
\examples{
nmea_meta(nmea_test_basic)

}
