% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-checksum.R
\name{nmea_parse_checksum}
\alias{nmea_parse_checksum}
\title{Calculate checksum metrics}
\usage{
nmea_parse_checksum(x)
}
\arguments{
\item{x}{An \code{\link[=nmea]{nmea()}} object.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns
\itemize{
\item \code{calc}: integer representation of the calculated checksum. This is
never \code{NA}.
\item \code{found}: integer representation of the checksum embedded in the
NMEA sentence or \code{NA} if the sentence did not contain a checksum
or contained one composed of invalid hex characters.
\item \code{start}: The zero-based index of the first character used to
calculate the checksum or \code{NA} if the starting character "$" was
not found.
\item \code{end}: The last character used to calculate the checksum such that
\code{end - start} is the number of characters used. \code{NA} if the
checksum character "*" was never found.
}

Both \code{start} and \code{end} are intended to be the values needed
to pass to \code{nmea_sub()} to extract the non-checksum components.
}
\description{
Calculate checksum metrics
}
\examples{
nmea_parse_checksum(nmea_test_basic)

# partial results for invalid nmea
nmea_parse_checksum("$GPGSA,A,1,,,")

}
