% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmearead.R
\name{nmea_parsetext}
\alias{nmea_parsetext}
\title{Parse NMEA text}
\usage{
nmea_parsetext(nmeatextlines)
}
\arguments{
\item{nmeatextlines}{The vector of strings containing NMEA lines.}
}
\value{
Data.table with all related NMEA information, with each row containing the GPS
details for the single timestamp.
}
\description{
Parse a vector of strings containing NMEA lines. Each vector element is a
new NMEA line.
}
