% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmearead.R
\name{nmea_parsetext_multifile}
\alias{nmea_parsetext_multifile}
\title{Parse NMEA text}
\usage{
nmea_parsetext_multifile(nmeatextlines, previousfirstlines = NULL)
}
\arguments{
\item{nmeatextlines}{The vector of strings containing NMEA lines.}

\item{previouslines}{Data.table rows returned from a previous call to this function,
to prepend to the current call. Use the \code{$firstlines} named item. See the example 
for how this works.}
}
\value{
Named list with the GPS data.table (\code{logtable}) and a data.table with the
last rows of the data.table that should be prepended to a next one 
(\code{firstlines}), or \code{NULL} if the log did not contain a full set at all.
}
\description{
Parse a vector of strings containing NMEA lines. Each vector element is a
new NMEA line. This function is used in loops where more than one file needs
to be appended.
}
\examples{
\dontrun{
prependlines = NULL
gpstable = NULL
for (file in filelist) {
  nmealines = readLines(file)
  info = nmea_parsetext_multifile(file,prependlines)
  prependlines = info$firstlines
  gpstable = rbind(gpstable,info$logtable)
}
}
}
