% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmearead.R
\name{nmea_readfiles}
\alias{nmea_readfiles}
\title{Read NMEA files}
\usage{
nmea_readfiles(pathtofiles)
}
\arguments{
\item{pathtofiles}{The full paths to the files.}
}
\value{
Data.table with all related NMEA information, with each row containing the GPS
details for the single timestamp.
}
\description{
Read NMEA compliant files. If more than one file path is provided, 
they are considered continuous and will be appended in the order provided.
}
