% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmearead.R
\name{nmea_tablestep2}
\alias{nmea_tablestep2}
\title{Execute step 2 in getting to a table.}
\usage{
nmea_tablestep2(logtable)
}
\arguments{
\item{logtable}{The first step NMEA data.table as returned from \code{\link{nmea_tablestep1}}.}
}
\value{
The final data.table with GPS details.
}
\description{
Internal function, do not call directly.
Step 2 transforms the data.table with \code{TYPE} and the other columns to the final resulting data.table.
}
