% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmea_talkers}
\alias{nmea_talkers}
\alias{nmea_message_types}
\title{NMEA talkers and message types}
\format{
An object of class \code{data.frame} with 48 rows and 2 columns.

An object of class \code{data.frame} with 67 rows and 2 columns.
}
\usage{
nmea_talkers

nmea_message_types
}
\description{
NMEA talkers and message types
}
\examples{
head(nmea_talkers)
head(nmea_message_types)

}
\keyword{datasets}
