% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmea_test_basic}
\alias{nmea_test_basic}
\title{Test NMEA vectors}
\format{
An object of class \code{character} of length 9.
}
\usage{
nmea_test_basic
}
\description{
Test NMEA vectors
}
\examples{
head(nmea_test_basic)

}
\keyword{datasets}
