% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-nmea.R
\name{read_nmea}
\alias{read_nmea}
\title{Extract NMEA sentences}
\usage{
read_nmea(
  x,
  ...,
  sentence_start = c("$", "!"),
  sentence_end = "\\n",
  max_length = 82L
)
}
\arguments{
\item{x}{An object from which to extract NMEA sentences. For the default
method this can be a \code{raw()} vector, a filename, a URL, or
a connection.}

\item{...}{Unused}

\item{sentence_start}{A vector of possible characters that denote
the start of a sentence. These must be a single character.}

\item{sentence_end}{A sequence that denotes the
end of a sentence. The NMEA specification suggests that sentences
always end with a carriage return (\\r) and a new line (\\n);
however, because these files are opened and saved on various systems,
a often only a newline follows a sentence.}

\item{max_length}{The maximum number of characters to scan after
and including \code{sentence_start} before giving up on
finding \code{sentence_end}. The NMEA specification suggest the maximum length
is 82 characters, however extensions occasionally send longer
sentences. This values includes both \code{sentence_start} and \code{sentence_end}.}
}
\value{
An \code{\link[=nmea]{nmea()}} vector.
}
\description{
For well-formed NMEA, existing tools for reading text are likely more
useful
}
\examples{
file <- system.file("extdata/basic.nmea", package = "nmea")
read_nmea(file)

}
