% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-package.R
\docType{package}
\name{nmea-package}
\title{nmea: Parse 'NMEA' Sentences}
\description{
Parse streams of bytes encoding National Marine Electronics
  Association ('NMEA') sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/nmea}
  \item Report bugs at \url{https://github.com/paleolimbot/nmea/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
