% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-spec.R
\name{nmea_col_character}
\alias{nmea_col_character}
\alias{nmea_col_double}
\alias{nmea_col_skip}
\alias{nmea_col_blob}
\alias{nmea_col_parse}
\alias{nmea_col_parse.nmea_col_character}
\alias{nmea_col_parse.nmea_col_skip}
\title{Specify NMEA field formats}
\usage{
nmea_col_character()

nmea_col_double()

nmea_col_skip()

nmea_col_blob()

nmea_col_parse(x, value)

\method{nmea_col_parse}{nmea_col_character}(x, value)

\method{nmea_col_parse}{nmea_col_skip}(x, value)
}
\arguments{
\item{x}{An \code{\link[=nmea]{nmea()}} object.}

\item{value}{An \code{\link[=nmea]{nmea()}} vector derived from \code{\link[=nmea_split_fields]{nmea_split_fields()}}.}
}
\value{
An object of class 'nmea_col'
}
\description{
Specify NMEA field formats
}
