% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmea-spec.R
\name{nmea_spec}
\alias{nmea_spec}
\title{Specify sentence parse structures}
\usage{
nmea_spec(..., .default = nmea_col_character())
}
\arguments{
\item{..., .default}{\code{\link[=nmea_col_character]{nmea_col_character()}} or similar.
These are most useful when named but this is not required.}
}
\value{
An object of class 'nmea_spec'.
}
\description{
Specify sentence parse structures
}
