% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-nmea.R
\name{read_nmea}
\alias{read_nmea}
\alias{read_nmea.default}
\title{Extract NMEA sentences}
\usage{
read_nmea(x, ...)

\method{read_nmea}{default}(
  x,
  ...,
  sentence_start = c("$", "!"),
  sentence_end = c("\\r\\n", "\\n"),
  max_length = 82L
)
}
\arguments{
\item{x}{An object from which to extract NMEA sentences}

\item{...}{Unused}

\item{sentence_start}{A vector of possible sequences that denote
the start of a sentence.}

\item{sentence_end}{A vector of possible sequences that denote the
end of a sentence. The NMEA specification suggests that sentences
always end with a carriage return (\\r) and a new line (\\n);
however, because these files are opened and saved on various systems,
a often only a newline follows a sentence.}

\item{max_length}{The maximum number of characters to scan after
and including \code{sentence_start} before giving up on
finding \code{sentence_end}. The NMEA specification suggest the maximum length
is 82 characters, however extensions occasionally send longer
sentences. This values inludes both \code{sentence_start} and \code{sentence_end}.}
}
\value{
An \code{\link[=nmea]{nmea()}} vector.
}
\description{
For well-formed NMEA, existing tools for reading text are likely more
useful
}
